+++ dart/lib/afhe/ciphertext.dart
@@ -37,7 +37,6 @@ final _LoadCiphertext _c_load_ciphertext = dylib
     .lookup<NativeFunction<_LoadCiphertextC>>('load_ciphertext')
     .asFunction();
 
-
 /// Represents an underlying C ciphertext object.
 ///
 /// A ciphertext is an encrypted message. It is the output of the encryption function,
@@ -48,6 +47,7 @@ class Ciphertext {
   // String text = "";
   /// Selects the [Backend] used for the encryption scheme.
   Backend backend = Backend();
+
   /// A pointer to the memory address of the underlying C++ object.
   Pointer obj = nullptr;
 
@@ -68,6 +68,18 @@ class Ciphertext {
   /// Saves the [Ciphertext] to a non-human-readable format.
   /// Useful for saving to disk or sending over the network.
   Pointer<Uint8> save() => _c_save_ciphertext(obj);
+
+  /// Converts a [Ciphertext] into a serialized binary format.
+  Uint8List toBytes() => save().asTypedList(saveSize);
+
+  /// Loads a [Ciphertext] from a serialized binary format.
+  Ciphertext.fromBytes(Afhe fhe, Uint8List bytes) {
+    final Pointer<Uint8> pointer = malloc.allocate<Uint8>(bytes.length);
+    // Copy the bytes to the allocated memory
+    final Uint8List pointerContent = pointer.asTypedList(bytes.length);
+    pointerContent.setAll(0, bytes);
+    obj = _c_load_ciphertext(fhe.library, pointer, bytes.length);
+  }
 }
 
 // --- noise budget ---