+++ dart/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.12
+
+- Added Ciphertext.fromBytes(Uint8List) Constructor
+- Added Ciphertext.toBytes() returns Uint8List
+- Added afhe.decodeDouble returns first element of decoded vector
+
+## 0.0.11
+
+Conan Release Management
+
 ## 0.0.8
 
 Auto-build Android binary in github releases